requirements.txt
python-dotenv==1.0.0
numpy==1.25.2
scikit-learn==1.3.0
scipy==1.11.2
matplotlib==3.7.2
pandas==2.1.0
requests==2.31.0